requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.4
cycler==0.10.0
Flask==2.0.3
grpcio==1.44.0
h11==0.13.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
keras==2.6.0
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.2.0
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
prettytable==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-graph-core==1.8.2
python-multipart==0.0.5
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.10.2
torchsummary==1.5.1
torchvision==0.11.3
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.17.5
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.7.0
